# internal/http/model_gen.go
// Package http provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen/v2 version v2.1.0 DO NOT EDIT.
package http

// Category defines model for Category.
type Category struct {
	Id   *int64 `json:"id,omitempty"`
	Name string `json:"name"`
}

// Pet defines model for Pet.
type Pet struct {
	Category Category `json:"category"`
	Id       *int64   `json:"id,omitempty"`
	Name     string   `json:"name"`
}

// Problem defines model for Problem.
type Problem struct {
	// Detail A human-readable explanation specific to this occurrence of the problem.
	Detail string `json:"detail"`

	// Instance A URI reference that identifies the specific occurrence of the problem. It may or may not yield further information if dereferenced.
	Instance string `json:"instance"`

	// Status The HTTP status code generated by the origin server for this occurrence of the problem.
	Status int `json:"status"`

	// Title A short, human-readable summary of the problem type.
	Title string `json:"title"`

	// Type A URI reference that identifies the problem type.
	Type *string `json:"type,omitempty"`
}

// GetPetsParams defines parameters for GetPets.
type GetPetsParams struct {
	// Category The category name to filter pets
	Category *[]string `form:"category,omitempty" json:"category,omitempty"`

	// From The id of the pet before starting to collect the result set
	From *int `form:"from,omitempty" json:"from,omitempty"`

	// Limit The numbers of items to return
	Limit *int `form:"limit,omitempty" json:"limit,omitempty"`
}

// AddPetJSONRequestBody defines body for AddPet for application/json ContentType.
type AddPetJSONRequestBody = Pet

// UpdatePetJSONRequestBody defines body for UpdatePet for application/json ContentType.
type UpdatePetJSONRequestBody = Pet